requirements.txt
pillow
django
django-nested-admin
psycopg2-binary
django-jazzmin
djangorestframework
django-phonenumber-field
phonenumbers
drf-yasg-json-api
drf-yasg
